+++ src/NaYeon/ImageSlider/components/style.ts
@@ -1,25 +1,20 @@
 import styled from "styled-components";
 
 export const St = {
-  Section: styled.section`
-    background-color: black;
-    height: 100vh;
-  `,
-
   MultipleSlider: styled.section`
     text-align: center;
     padding: 0 4rem;
   `,
 
   Title: styled.h1`
-    font-size: 2rem;
+    font-size: 3rem;
     font-weight: bold;
     color: white;
     padding: 3rem;
   `,
 
   ImageWrapper: styled.div`
-    height: 350px;
+    height: 550px;
     border: 20px solid black;
     box-sizing: border-box;
   `,
