+++ src/NaYeon/ImageSlider/core/temp.ts
