+++ src/NaYeon/ImageSlider/util/temp.ts
