+++ src/NaYeon/SliderBar/core/temp.ts
