+++ src/NaYeon/SliderBar/util/temp.ts
