--- src/NaYeon/image_slider/index.tsx
@@ -1,10 +1,0 @@
-import MultipleSlide from "./components/multiple";
-import { St } from "./style";
-
-export default function image_slider() {
-  return (
-    <St.Section>
-      <MultipleSlide />
-    </St.Section>
-  );
-}
