--- src/NaYeon/slider_bar/index.tsx
@@ -1,10 +1,0 @@
-import ImageSilder from "./components/slider";
-import { St } from "./style";
-
-export default function slider_bar() {
-  return (
-    <St.Section>
-      <ImageSilder />
-    </St.Section>
-  );
-}
